refactor: Accept CGFloat instead of Int in addDashBorder extension COOEE-116
Shape accepts value in CGFloat hence converted accepting CGFloat

+++ CooeeSDK/trigger/inapp/renderer/RendererExtensions.swift
@@ -16,7 +16,7 @@ extension UIView {
        - dashGap: With of gap between two dash of stroke in Int
        - cornerRadius: Radius of corner in Int
      */
-    func addDashedBorder(colour: UIColor, width: Int, dashWidth: Int, dashGap: Int, cornerRadius: Int) {
+    func addDashedBorder(colour: UIColor, width: CGFloat, dashWidth: CGFloat, dashGap: CGFloat, cornerRadius: CGFloat) {
         let shapeLayer: CAShapeLayer = CAShapeLayer()
         let frameSize = self.frame.size
         let shapeRect = CGRect(x: 0, y: 0, width: frameSize.width, height: frameSize.height)
@@ -25,9 +25,9 @@ extension UIView {
         shapeLayer.position = CGPoint(x: frameSize.width / 2, y: frameSize.height / 2)
         shapeLayer.fillColor = UIColor.clear.cgColor
         shapeLayer.strokeColor = colour.cgColor
-        shapeLayer.lineWidth = CGFloat(width)
+        shapeLayer.lineWidth = width
         shapeLayer.lineJoin = CAShapeLayerLineJoin.round
-        shapeLayer.cornerRadius = CGFloat(cornerRadius)
+        shapeLayer.cornerRadius = cornerRadius
         shapeLayer.lineDashPattern = [NSNumber(value: dashWidth), NSNumber(value: dashGap)]
         shapeLayer.path = UIBezierPath(roundedRect: shapeRect, cornerRadius: 5).cgPath
